Fix JsonRepositoryLoaderTest
