--- tests/fixtures/repositories/invalid/broken-url.json
@@ -1,11 +1,0 @@
-{
-  "phars": {
-    "foobar": {
-      "url": "123",
-      "checksum": {
-        "type": "sha-256",
-        "value": "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef"
-      }
-    }
-  }
-}
